Propagator: hide EigenStepperError implementation
