+++ Core/src/Propagator/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources_local(
   ActsCore
   PRIVATE
+    EigenStepperError.cpp
     StraightLineStepper.cpp
     detail/PointwiseMaterialInteraction.cpp
     detail/CovarianceEngine.cpp
